chore/#24: ResponseAccountState 파일 분리
