+++ app/src/main/java/org/sopt/starbanking/data/dto/response/ResponseAccountState.kt
@@ -14,41 +14,3 @@ data class ResponseAccountState(
     @SerialName("contractPeriod")
     val contractPeriod: Int
 )
-
-@Serializable
-data class ResponseSavingsState(
-    @SerialName("savingAccountName")
-    val savingAccountName: String,
-    @SerialName("accountNumber")
-    val accountNumber: String,
-    @SerialName("totalBalance")
-    val totalBalance: Int,
-    @SerialName("startDate")
-    val startDate: String,
-    @SerialName("endDate")
-    val endDate: String,
-    @SerialName("dDay")
-    val dDay: String,
-    @SerialName("preferentialRate")
-    val preferentialRate: String,
-    @SerialName("maxAppliedRate")
-    val maxAppliedRate: String,
-    @SerialName("deposits")
-    val deposits: List<DepositDto>
-)
-
-@Serializable
-data class DepositDto(
-    @SerialName("id")
-    val id: Long,
-    @SerialName("count")
-    val count: Int,
-    @SerialName("depositDate")
-    val depositDate: String,
-    @SerialName("payment")
-    val payment: Int,
-    @SerialName("afterPaymentBalance")
-    val afterPaymentBalance: Int,
-    @SerialName("appliedRate")
-    val appliedRate: String
-)
